neat: making var types consistent in test_nn.cpp

+++ src/test_nn.cpp
@@ -3,8 +3,8 @@
 #include <uxhw.h>
 
 // Function to perform forward pass in a single layer model
-double forwardPass(const std::vector<double>& inputs, const std::vector<double>& weights) {
-    double output = 0.0;
+float forwardPass(const std::vector<float>& inputs, const std::vector<float>& weights) {
+    float output = 0.0f;
     int input_size = inputs.size();
     for (int j = 0; j < input_size; j++) {
         output += inputs[j] * weights[j];
@@ -18,11 +18,11 @@ int main() {
     float w2 = -UxHwFloatUniformDist(0.4f, 0.6f);
     float w3 = UxHwFloatUniformDist(0.2f, 0.4f);
     
-    std::vector<double> weights = {w1, w2, w3};
-    std::vector<double> inputs = {1.0, 2.0, 3.0};
+    std::vector<float> weights = {w1, w2, w3};
+    std::vector<float> inputs = {1.0f, 2.0f, 3.0f};
 
     // Calculate output
-    double output = forwardPass(inputs, weights);
+    float output = forwardPass(inputs, weights);
 
     // Print the output
     std::cout << "Output: " << output << std::endl;
